71 game env design
